requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
deepface==0.0.99
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fsspec==2026.2.0
gast==0.7.0
gdown==6.0.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==25.3.0
h5py==3.14.0
idna==3.11
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.5
lightphe==0.0.23
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
pillow==12.1.1
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyWavelets==1.9.0
rawpy==0.26.1
referencing==0.37.0
requests==2.33.1
retina-face==0.0.17
rich==15.0.0
rpds-py==0.30.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.56.0
sympy==1.13.1
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
toml==0.10.2
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.8
wrapt==2.1.2
